% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_vectorization_internal.R
\name{assemble_contact_matrix}
\alias{assemble_contact_matrix}
\title{Assembles different parts of contact matrix, also accounting for location-specific sd}
\usage{
assemble_contact_matrix(pars)
}
\arguments{
\item{pars}{parameters}
}
\value{
contact matrix
}
\description{
Assembles different parts of contact matrix, also accounting for location-specific sd
}
