% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_vectorization_internal.R
\name{augment_temporal}
\alias{augment_temporal}
\title{Augment temporal}
\usage{
augment_temporal(params_temporal, temp_list, param_name)
}
\arguments{
\item{params_temporal}{temporal parameters}

\item{temp_list}{temporal parameter list}

\item{param_name}{parameter name}
}
\value{
temporal list with additional parameters added
}
\description{
Augment temporal
}
