% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_data_stats.R
\name{calc_data_stats}
\alias{calc_data_stats}
\title{Calculates metrics for calibration from the data}
\usage{
calc_data_stats(
  data_actual,
  dates,
  rescale_factors = list(cases = rep(1, 4), deaths = rep(1, 4), hosp = rep(1, 4),
    negtests = rep(1, 4)),
  stats_to_include = c("cases", "deaths", "hosp")
)
}
\arguments{
\item{data_actual}{the data to use}

\item{dates}{which dates to include}

\item{rescale_factors}{i.e. the mean of each time series to normalize by}

\item{stats_to_include}{which of cases, deaths and hosp to include}
}
\value{
vector of statistics
}
\description{
Calculates metrics for calibration from the data
}
