% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_model_stats.R
\name{calc_model_stats}
\alias{calc_model_stats}
\title{Runs model and alculates metrics for calibration from the model output}
\usage{
calc_model_stats(
  pars,
  pars_names,
  pars_base,
  pars_temporal,
  state,
  start_from_first_inf = FALSE,
  start_date = NULL,
  end_date = NULL,
  dates,
  rescale_factors = list(cases = rep(1, 4), deaths = rep(1, 4), hosp = rep(1, 4),
    negtests = rep(1, 4)),
  stats_to_include = c("cases", "deaths", "hosp")
)
}
\arguments{
\item{pars}{vector of parameters (i.e. those being calibrated)}

\item{pars_names}{names of pars parameters}

\item{pars_base}{other fixed parameters}

\item{pars_temporal}{other temporal parameters}

\item{state}{initial state}

\item{start_from_first_inf}{whether to start from very beginning}

\item{start_date}{if start_from_first_inf==TRUE, start from first_inf_day, if NULL, use min(dates)}

\item{end_date}{if NULL, use max(dates)}

\item{dates}{which dates to include}

\item{rescale_factors}{i.e. the mean of each time series to normalize by}

\item{stats_to_include}{which of cases, deaths and hosp to include}
}
\value{
vector of statistics
}
\description{
Runs model and alculates metrics for calibration from the model output
}
