% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sse_multi.R, R/calc_sse_simple.R
\name{calc_sse_multi}
\alias{calc_sse_multi}
\alias{calc_sse_simple}
\title{Calculate error between model and data}
\usage{
calc_sse_multi(
  pars,
  pars_names,
  pars_base,
  pars_temporal,
  state,
  data_actual,
  start_from_first_inf = FALSE,
  start_date = NULL,
  end_date = NULL
)

calc_sse_simple(
  pars,
  pars_names,
  pars_base,
  pars_temporal,
  state,
  data_actual,
  start_from_first_inf = FALSE,
  start_date = NULL,
  end_date = NULL
)
}
\arguments{
\item{pars}{the vector of parameters from the optimization routine, and age-specific parameters should have _age
appended to the parameter name (e.g. sd_1)}

\item{pars_names}{the names of the parameters}

\item{pars_base}{the baseline set of parameters to which pars are applied}

\item{pars_temporal}{any parameters that need to be updated through time, specified as a list of lists, 
where each list has a date=<calendar date> and the parameter name-value pairs}

\item{state}{the model state at the start, either from a previous run or generated from initial_state()}

\item{data_actual}{the data to compare to}

\item{start_from_first_inf}{(optional) if the model should start from the first infection 
rather than the start of the data}

\item{start_date}{(optional) the date to start if not the start of data or first infection}

\item{end_date}{(optional) the date to end if not the end of data}
}
\value{
vector of errors for cases, deaths, hospitalizations, and tests
}
\description{
This functions runs the model with the specified parameters and compares the results to data. It is designed to 
be called from an optimizer, thus parameters are specified in several ways, see below.
}
