% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dynamic_sd.R
\name{calculate_dynamic_sd}
\alias{calculate_dynamic_sd}
\alias{calculate_dynamic_sd_cases}
\alias{calculate_dynamic_sd_hosp}
\alias{calculate_dynamic_sd_cases_and_hosp}
\title{Functions used to determine dynamic sd activation or relaxation}
\usage{
calculate_dynamic_sd_cases(pars, sd_update_metrics)

calculate_dynamic_sd_hosp(pars, sd_update_metrics)

calculate_dynamic_sd_cases_and_hosp(pars, sd_update_metrics)
}
\arguments{
\item{pars}{model parameters}

\item{sd_update_metrics}{metrics from model}
}
\value{
new sd value
}
\description{
Use one of the three following functions should be specified for the "dynamic_sd_func" parameter
\itemize{
\item calculate_dynamic_sd_cases on a date that we want to update the dynamic sd, calculate what it should be (it might not change)
\item calculate_dynamic_sd_hosp # uses hospitalizations to evaluate lock and release thresholds
\item calculate_dynamic_sd_cases_and_hosp uses cases OR hospitalizations to lock and cases AND hospitalizations to release. 
}
}
