% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_sd_internal.R
\name{calculate_sd_metrics}
\alias{calculate_sd_metrics}
\title{calculates epidemic metrics used to calculate the dynamic sd}
\usage{
calculate_sd_metrics(model_out, parameters)
}
\arguments{
\item{model_out}{output metrix at timepoint to calculate metrics}

\item{parameters}{parameters used to run model}
}
\value{
metrics
}
\description{
calculates epidemic metrics used to calculate the dynamic sd
}
