% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_eq_internal.R
\name{calculate_vaccine_distribution}
\alias{calculate_vaccine_distribution}
\title{Calculate vaccine distribution for S and R vectors for this time step}
\usage{
calculate_vaccine_distribution(
  S,
  N,
  V_vac_prop,
  n_age,
  n_strain,
  n_vaccine,
  n_recovered,
  vac_total,
  vac_dist,
  vac_priority,
  vac_coverage
)
}
\arguments{
\item{S}{The current vector for the S box}

\item{N}{The current vector for the N box, i.e. age distribution of everyone living}

\item{V_vac_prop}{Matrix that describes how to distribute vaccines within susceptible and recovered parts of S, each column should sum to 0 (neg to remove, pos to add, also gives propotion doses per vaccine)}

\item{n_age}{Number of age groups}

\item{n_strain}{Number of strains}

\item{n_vaccine}{Number of vaccines (1 = no vaccine)}

\item{n_recovered}{Number of recovered states}

\item{vac_total}{Number of doses to distribute (per vaccine #Mia note: combine vac_num and vac_dist)total across all vaccines)}

\item{vac_dist}{Percentage to distribute to each age group}

\item{vac_priority}{Vector of age groups in priority order to distribute doses once groups saturate doses #Mia note: Please Remove}

\item{vac_coverage}{What percentage of each age group to vaccinate before considering saturated #Mia note: This could be handled by subdividing population?}
}
\value{
list with a several vectors: S_vaccines: vector to add to S to redistribute from unvaccinated box to
the different vaccinated boxes, R_vaccines: vector to add to R to redistribute from unvaccinated box to
the different vaccinated boxes also by strain, total_vaccinated: the total number vaccinated by age group, so the
rest of the vector (the spots for strains and vaccines) are zero
}
\description{
Calculated to vectors to add to S and R to redistribute amongst the boxes in order to vaccinate. The tricky part
is that we need to not exceed the intended coverage rate and also not allow any box to go negative. However, we need
to spit the available doses by age group but also by S and R (and by strain for R). Further, even if we haven't yet 
hit coverage limits, there may not be enough individuals to vaccinate because some may be in varaious infected boxes
and not yet available in the R box.
}
