% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_vectorization_internal.R
\name{create_block}
\alias{create_block}
\title{Generalized creation for W and Y matrices}
\usage{
create_block(M, n_age)
}
\arguments{
\item{M}{base matrix}

\item{n_age}{number of age groups}
}
\value{
matrix
}
\description{
Generalized creation for W and Y matrices
}
