% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_vectorization_internal.R
\name{create_block_diag}
\alias{create_block_diag}
\title{creates a diagonal matrix with diag_matrix on the diagonal and zero_matrix everywhere else, with dim x dim blocks}
\usage{
create_block_diag(diag_block, zero_block, dim)
}
\arguments{
\item{diag_block}{block for diagonal}

\item{zero_block}{block for everywhere else}

\item{dim}{how many blocks}
}
\value{
contructed matrix
}
\description{
creates a diagonal matrix with diag_matrix on the diagonal and zero_matrix everywhere else, with dim x dim blocks
}
