% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_vectorization_internal.R
\name{create_conv_by_replicate}
\alias{create_conv_by_replicate}
\title{Creates M that converts between S and other boxes that are by strain}
\usage{
create_conv_by_replicate(n_age, n_strain, n_vaccine_eff)
}
\arguments{
\item{n_age}{number of age groups}

\item{n_strain}{number of strains}

\item{n_vaccine_eff}{number of vaccine and recovered states}
}
\value{
M matrix
}
\description{
M %*% S will be the dim for E, I etc.
and t(M) %*% param_vec will be the dim for S
}
