% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_vectorization_internal.R
\name{create_conv_by_zeros}
\alias{create_conv_by_zeros}
\title{Creates an alternative M that converts between S and other boxes that are by strain}
\usage{
create_conv_by_zeros(n_age, n_strain, n_vaccine_eff)
}
\arguments{
\item{n_age}{number of age groups}

\item{n_strain}{number of strains}

\item{n_vaccine_eff}{number of vaccine and recovered states}
}
\value{
M matrix with 0's for non-S compartments
}
\description{
in this case the additional dimensions are filled with 0's
this avoids double (or triple) counting for calculating things like negative tests
}
