% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_vectorization_internal.R
\name{create_pop_conv_full}
\alias{create_pop_conv_full}
\title{Creates matrix to sum and replicate population across boxes, keeping it age structured}
\usage{
create_pop_conv_full(n_age, n_strain, n_vaccine_eff)
}
\arguments{
\item{n_age}{number of age groups}

\item{n_strain}{number of strains}

\item{n_vaccine_eff}{number of vaccine and recovered states}
}
\value{
matrix
}
\description{
So it will sum across strains and vaccines, but replicate the final 
age-structured population across strains and vaccines. Used to calculate lambda.
}
