% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_Q.R
\name{create_Q}
\alias{create_Q}
\title{Generalized Q matrix creation}
\usage{
create_Q(immune.list, n_age, temporal = FALSE)
}
\arguments{
\item{immune.list}{a nested list, see description}

\item{n_age}{number of age groups}

\item{temporal}{TRUE if temporal}
}
\value{
Q matrix
}
\description{
Generalized Q matrix creation input type has structure immune.list
immune.list must be a nested list. 
Each sublist represents an immune state and must contain an element "Coverage"
"Coverage" is a vector that must have an entry for each immune state listed. Those entries
 are the contribution towards the coverage "felt" by individuals in the immune state
 So an entry: immune.list$Recovered$Coverage = c("Naive" = 0, "Recovered" = 0, "Vax" = 1)
 means that individuals in the recovered class compute their current coverage by looking at individuals in the "Vax" class
 Note that the above also implies that there are only three immune states: "Naive", "Recovered", and "Vax"
}
\details{
Note that n_age is the size of the block diagonals (will typically represent age groups)
}
