% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_V_recovered_by_vax.R
\name{create_V_recovered_by_vax}
\alias{create_V_recovered_by_vax}
\title{Create V matrix for recovered states by vaccination status}
\usage{
create_V_recovered_by_vax(
  n_age,
  n_vaccine,
  vac_prop = rep(1/(n_vaccine - 1), n_vaccine - 1)
)
}
\arguments{
\item{n_age}{number of age groups}

\item{n_vaccine}{number of vaccines}

\item{vac_prop}{how to proportionally distribute vaccines}
}
\value{
V matrix
}
\description{
Recovered from infection goes into either recovered non vaccinated, or according to vaccine, but not by strain
}
