% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_V_single_recovered.R
\name{create_V_single_recovered}
\alias{create_V_single_recovered}
\title{Create V matrix for a single recovered box}
\usage{
create_V_single_recovered(
  n_age,
  n_vaccine,
  vac_prop = rep(1/(n_vaccine - 1), n_vaccine - 1)
)
}
\arguments{
\item{n_age}{number of age groups}

\item{n_vaccine}{number of vaccines}

\item{vac_prop}{how to proportionally distribute vaccines}
}
\value{
Y matrix
}
\description{
Recovery from infection goes into susceptible for single recovered compartment no matter strain
}
