% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_W.R
\name{create_W}
\alias{create_W}
\title{Generalized W matrix creation}
\usage{
create_W(immune.list, n_age, temporal = FALSE)
}
\arguments{
\item{immune.list}{a nested list, see description}

\item{n_age}{number of age groups}

\item{temporal}{TRUE if temporal}
}
\value{
W matrix
}
\description{
Generalized W matrix creation input type has structure immune.list
immune.list must be a nested list. 
Each sublist represents an immune state and must contain an element "Waning"
"Waning" is a vector that must have an entry for each immune state listed. Those entries
 are the rate at which one immune state decays into the other
 So an entry: immune.list$Recovered$Waning = c("Naive" = .1, "Recovered = 0)
 means that individuals travel Recovered->Naive at a rate of 0.1/day.
 Note that the above also implies that there are only two immune states: "Naive" and "Recovered"
}
\details{
Note that n_age is the size of the block diagonals (will typically represent age groups)
}
