% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_W_most_recent.R
\name{create_W_most_recent}
\alias{create_W_most_recent}
\title{W with vaccine and natural infection waning}
\usage{
create_W_most_recent(n_age, n_vaccine, n_recovered, waning_matrix)
}
\arguments{
\item{n_age}{number of age groups}

\item{n_vaccine}{number of vaccines}

\item{n_recovered}{number of recovered groups}

\item{waning_matrix}{a (m + p) x n matrix giving the waning per age for each vaccine and strain infection}
}
\value{
W matrix
}
\description{
that is columns for age and rows for vaccines and strains
note the first vaccine is 'no vaccine' so the first row should be all 0 (like VE)
}
