% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_W_no_waning.R
\name{create_W_no_waning}
\alias{create_W_no_waning}
\title{W with no waning}
\usage{
create_W_no_waning(n_age, n_vaccine, n_recovered)
}
\arguments{
\item{n_age}{number of age groups}

\item{n_vaccine}{number of vaccines}

\item{n_recovered}{number of recovered groups}
}
\value{
W matrix
}
\description{
no waning, like in original version of model, protection from infection lasts forever
}
