% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_Y.R
\name{create_Y}
\alias{create_Y}
\title{Generalized Y matrix creation}
\usage{
create_Y(immune.list, strain.list, n_age)
}
\arguments{
\item{immune.list}{a nested list}

\item{strain.list}{a nested list}

\item{n_age}{number of age groups}
}
\value{
Y matrix
}
\description{
Note that n_age is the size of the block diagonals (will typically represent age groups)
}
