% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_Y_recovered_by_strain.R
\name{create_Y_recovered_by_strain}
\alias{create_Y_recovered_by_strain}
\title{Create Y matrix for recovered compartment for each strain}
\usage{
create_Y_recovered_by_strain(n_age, n_vaccine, n_strain)
}
\arguments{
\item{n_age}{number of age groups}

\item{n_vaccine}{number of vaccines}

\item{n_strain}{number of strains}
}
\value{
Y matrix
}
\description{
# recovery from infection goes into susceptible for appropriate strain
}
