% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_Y_recovered_by_vax.R
\name{create_Y_recovered_by_vax}
\alias{create_Y_recovered_by_vax}
\title{Create Y matrix for recovered states by vaccination status}
\usage{
create_Y_recovered_by_vax(n_age, n_vaccine, n_strain)
}
\arguments{
\item{n_age}{number of age groups}

\item{n_vaccine}{number of vaccines}

\item{n_strain}{number of strains}
}
\value{
Y matrix
}
\description{
Recovered from infection goes into either recovered non vaccinated, or according to vaccine, but not by strain
}
