% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_Y_single_recovered.R
\name{create_Y_single_recovered}
\alias{create_Y_single_recovered}
\title{Create Y matrix for a single recovered box}
\usage{
create_Y_single_recovered(
  n_age,
  n_vaccine,
  n_strain,
  maintain_vax_staus = TRUE
)
}
\arguments{
\item{n_age}{number of age groups}

\item{n_vaccine}{number of vaccines}

\item{n_strain}{number of strains}

\item{maintain_vax_staus}{TRUE for recovered vaccinated to go to vaccnated compartment, FALSE to recovered compartment}
}
\value{
Y matrix
}
\description{
Recovery from infection goes into susceptible for single recovered compartment no matter strain
}
