% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_data_internal.R
\name{diff_col}
\alias{diff_col}
\title{helper function to go from cumulative to daily counts}
\usage{
diff_col(x)
}
\arguments{
\item{x}{vector of cumulative}
}
\value{
daily values
}
\description{
helper function to go from cumulative to daily counts
}
