% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_vectorization_internal.R
\name{expand_contact_matrix}
\alias{expand_contact_matrix}
\title{Expand contact matrix for multiple strains and vaccines}
\usage{
expand_contact_matrix(c, n_strain, n_vaccine_eff)
}
\arguments{
\item{c}{contact matrix}

\item{n_strain}{number of strains}

\item{n_vaccine_eff}{number of vaccines}
}
\value{
expanded contact matrix
}
\description{
Expand contact matrix for multiple strains and vaccines
}
