% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_vectorization_internal.R
\name{expand_strain_import}
\alias{expand_strain_import}
\title{Expands that strain import so that 0's are added for vaccine entries}
\usage{
expand_strain_import(strain_import_matrix, n_age, n_vaccine_eff)
}
\arguments{
\item{strain_import_matrix}{strain imports per day}

\item{n_age}{number of age groups}

\item{n_vaccine_eff}{how many vaccine and recovered states}
}
\value{
expanded strain import vector
}
\description{
result is a n x p x m vector
strain_import_matrix is a m x n matrix
}
