% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_vectorization_internal.R
\name{expand_vector}
\alias{expand_vector}
\title{Expand a vector of age specific parameters}
\usage{
expand_vector(v, n_strain, n_vaccine, n_recovered)
}
\arguments{
\item{v}{vector to expand}

\item{n_strain}{number of strains}

\item{n_vaccine}{number of vaccines}

\item{n_recovered}{number of recovered states}
}
\value{
expanded vector
}
\description{
Expand a vector of age specific parameters
}
