% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_vectorization_internal.R
\name{expand_vector_params}
\alias{expand_vector_params}
\title{Expands vectorized parameters}
\usage{
expand_vector_params(pars)
}
\arguments{
\item{pars}{parameters pre-expansion}
}
\value{
parameters expanded as well as matrices added
}
\description{
Rewrites parameter vectors to be the necessary length for vaccines and strains
}
