% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_vectorization_internal.R
\name{expand_vector_vaceff}
\alias{expand_vector_vaceff}
\title{Expand a vector of age specific parameters by a set of vaccine efficacies}
\usage{
expand_vector_vaceff(v, vac_eff)
}
\arguments{
\item{v}{vector to expand}

\item{vac_eff}{matrix of vaccine efficacies with rows for each strain}
}
\value{
expanded vector
}
\description{
NOTE: the first vac_eff should be 0 to correspond to no vaccine
}
