% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_final_state.R
\name{extract_final_state}
\alias{extract_final_state}
\title{extracts the final state and the end of model output}
\usage{
extract_final_state(model_out)
}
\arguments{
\item{model_out}{matrix model output}
}
\value{
state vector
}
\description{
suitable to using to start the next model run
(i.e. with out the time or sd columns)
}
