% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_internal.R
\name{extract_state}
\alias{extract_state}
\title{extracts just the state variables that can be used to start running model from that state or to calculate Reff}
\usage{
extract_state(model_out, row_idx)
}
\arguments{
\item{model_out}{matrix model output}

\item{row_idx}{which row}
}
\value{
state vector
}
\description{
extracts just the state variables that can be used to start running model from that state or to calculate Reff
}
