% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_non_compartments.R
\name{filter_non_compartments}
\alias{filter_non_compartments}
\title{Filters out the  states that aren't actually compartments}
\usage{
filter_non_compartments(data_long)
}
\arguments{
\item{data_long}{model output in long format}
}
\value{
model data with only true model compartments, not tracking compartments
}
\description{
i.e, sd, R-eff, the various cum_xxx tracking compartments
 useful for summing value across vaccination or strain
}
