% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_runner_internal.R
\name{generate_initial_ramp}
\alias{generate_initial_ramp}
\title{Ramp to initial sd value}
\usage{
generate_initial_ramp(sd, start_date, start_ramp_date, end_ramp_date)
}
\arguments{
\item{sd}{initial sd value}

\item{start_date}{model start date}

\item{start_ramp_date}{when to start ramping sd}

\item{end_ramp_date}{when to attain initial sd value}
}
\value{
temporal parameters for sd ramp at epidemic start
}
\description{
Ramp to initial sd value
}
