% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_sd_internal.R
\name{generate_sd_dynamics}
\alias{generate_sd_dynamics}
\title{generate the set of dates we need to (potentially) adjust the dynamic sd and adds them to the temporal parameters}
\usage{
generate_sd_dynamics(pars_base, end_date)
}
\arguments{
\item{pars_base}{base parameters}

\item{end_date}{end of model run}
}
\value{
list of dates to potentially adjust sd
}
\description{
generate the set of dates we need to (potentially) adjust the dynamic sd and adds them to the temporal parameters
}
