% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_runner_internal.R
\name{generate_sd_ramp}
\alias{generate_sd_ramp}
\title{Ramps linearly between two sd values.}
\usage{
generate_sd_ramp(sd_initial, sd_final, start_ramp_date, end_ramp_date)
}
\arguments{
\item{sd_initial}{starting sd}

\item{sd_final}{ending sd}

\item{start_ramp_date}{date of initial sd}

\item{end_ramp_date}{date of final sd}
}
\value{
temporal parameters to ramp sd
}
\description{
Generates the sd ramping that used to be in the model code for the initial transition 
to sd at the very start of the epidemic or for the transition between 2 sd values. 
For now this is hard-coded to stairstep up weekly.
}
