% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_runner_internal.R
\name{generate_temporal_ramp}
\alias{generate_temporal_ramp}
\title{Generates ramps for subsequent sd parameters after initial lockdown, if any}
\usage{
generate_temporal_ramp(pars_temporal, start_ramp_date, start_sd)
}
\arguments{
\item{pars_temporal}{temporal parameters (with sd)}

\item{start_ramp_date}{date to start ramping}

\item{start_sd}{inital sd}
}
\value{
temporal parameters for sd ramps
}
\description{
Generates ramps for subsequent sd parameters after initial lockdown, if any
}
