% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_effective_internal.R
\name{Get.Trans.matrix}
\alias{Get.Trans.matrix}
\title{Get Trans matrix}
\usage{
Get.Trans.matrix(parameters, state, sd.mult.side = "both")
}
\arguments{
\item{parameters}{model parameters}

\item{state}{current state}

\item{sd.mult.side}{whether sd is applied to both sides of contact matrix}
}
\value{
matrix
}
\description{
Get Trans matrix
}
