% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_vectorization_internal.R
\name{get_attribute_from_specs}
\alias{get_attribute_from_specs}
\title{Get attribute from specs}
\usage{
get_attribute_from_specs(input.list, attribute, temporal = F)
}
\arguments{
\item{input.list}{input}

\item{attribute}{attribute}

\item{temporal}{TRUE if temporal}
}
\value{
attribute
}
\description{
Get attribute from specs
}
