% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_date_from_model_day.R
\name{get_date_from_model_day}
\alias{get_date_from_model_day}
\title{Gets date in calendar time from a day in model time}
\usage{
get_date_from_model_day(model_day, model_day0_date)
}
\arguments{
\item{model_day}{day to translate}

\item{model_day0_date}{calendar date for model day0}
}
\value{
calendar date
}
\description{
Gets date in calendar time from a day in model time
}
