% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_data_param_sets.R
\name{get_model_data_param_sets}
\alias{get_model_data_param_sets}
\title{Runs model for a matrix of parameterizations}
\usage{
get_model_data_param_sets(
  pars_matrix,
  params_names,
  params_base,
  params_temporal,
  init_state,
  kc_data = NULL,
  start_date = NULL,
  end_date = NULL,
  calc_r_eff = TRUE,
  out_type = c("wide", "long", "matrix"),
  parallel = FALSE
)
}
\arguments{
\item{pars_matrix}{matrix with each row a parameterization}

\item{params_names}{names of columns of pars_matrix}

\item{params_base}{base parameters}

\item{params_temporal}{temporal parameters}

\item{init_state}{inital state}

\item{kc_data}{(optional) data to set end date from}

\item{start_date}{(optional) start date, or will use first_inf_day}

\item{end_date}{(optional) end date}

\item{calc_r_eff}{whetether to also calculate Reff}

\item{out_type}{wide is a wide data frame, long is a long data frame, and matrix is the ode matrix output}

\item{parallel}{wheterh to run in parallel}
}
\value{
list of model output for each parameterization
}
\description{
get a list where each list item is a dataframe of model output
the option parallel is to run in parallel, this assumes a cluster has ben set up outside this function and foreach will use it
}
