% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_day_from_date.R
\name{get_model_day_from_date}
\alias{get_model_day_from_date}
\title{Gets day in model time from a date object from defined model day 0}
\usage{
get_model_day_from_date(date, model_day0_date)
}
\arguments{
\item{date}{date to translate}

\item{model_day0_date}{calendar date for model day0}
}
\value{
model day
}
\description{
Gets day in model time from a date object from defined model day 0
}
