% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameters_from_specs.R
\name{get_parameters_from_specs}
\alias{get_parameters_from_specs}
\title{Get parameters}
\usage{
get_parameters_from_specs(scenario.specs, params, params_temporal)
}
\arguments{
\item{scenario.specs}{scenarios}

\item{params}{base parameters}

\item{params_temporal}{temporal parameters}
}
\value{
parameters
}
\description{
Get parameters
}
