% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params.R
\name{get_params}
\alias{get_params}
\title{Gets parameters}
\usage{
get_params(pars, pars_names, pars_base)
}
\arguments{
\item{pars}{Vector of parameter, e.g. from the optimizer}

\item{pars_names}{Names of the parameter in pars}

\item{pars_base}{Base parameter values, to be updated with those in pars}
}
\value{
List of parameters suitable for calling model_eq
}
\description{
This is a helper function since the optimizer just uses an unnamed vector of parameters.
We pass an additional vector of the names of the parameters, that can be used to determine
what parameters are in the pars vector and this overwrite the correct parameter in the 
parameter list pars_base.
}
\details{
For age-varying parameters, because we can only pass a single vector to the optimizer,
these are passed as paramname_age, for example sd_1, which is parsed and the appropriate
parameter in the vector is updated.

With the vectorization of the model, it is important that get_params() is called before run_model(),
which will expand the age-specific parameters as necessary for strains and vaccines.
}
