% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_vectorization_internal.R
\name{get_simple_vectorized_params}
\alias{get_simple_vectorized_params}
\title{Gets parameters that are vectorized by age and thus need to be expanded for strains and vaccines}
\usage{
get_simple_vectorized_params()
}
\value{
vector of parameter names
}
\description{
Gets parameters that are vectorized by age and thus need to be expanded for strains and vaccines
}
