% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_starting_state.R
\name{get_starting_state}
\alias{get_starting_state}
\title{Gets staring state}
\usage{
get_starting_state(
  parameters,
  starting_conditions,
  immune_frac,
  strain_prevalence_frac
)
}
\arguments{
\item{parameters}{base model parameters}

\item{starting_conditions}{a list specifying total_prevalence, total_population, cumulative_incidence, and proportion_vaccinated}

\item{immune_frac}{a list specifying the proportions in each vaccine state by age group}

\item{strain_prevalence_frac}{a vertor for prevalence of each strain}
}
\value{
starting state vector
}
\description{
Gets staring state
}
