% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_starting_state_from_scenario.R
\name{get_starting_state_from_scenario}
\alias{get_starting_state_from_scenario}
\title{Gets starting state for scenario}
\usage{
get_starting_state_from_scenario(
  parameters,
  starting_conditions,
  scenario.specs
)
}
\arguments{
\item{parameters}{base model parameters}

\item{starting_conditions}{a list specifying total_prevalence, total_population, cumulative_incidence, and proportion_vaccinated}

\item{scenario.specs}{a list which contains a strains.list which is a list of streans, each specifying
rel.transmissability. rel.severity, import.rate, frac.prevalence, and frac.cum.inc}
}
\value{
starting state vector
}
\description{
Gets starting state for scenario
}
