% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_state_order.R
\name{get_state_order}
\alias{get_state_order}
\title{Model states}
\usage{
get_state_order()
}
\value{
vector of state names
}
\description{
note that matching to order here and in the model definition is critical and thus we define indices
unfortunately can't pass as list of vectors like parameters or ode will only allow returning 
results the same length as the list
}
