% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_temporal_params.R
\name{get_temporal_params}
\alias{get_temporal_params}
\title{Gets temporal parameters}
\usage{
get_temporal_params(pars, pars_names, pars_temporal_base)
}
\arguments{
\item{pars}{Vector of parameter, e.g. from the optimizer}

\item{pars_names}{Names of the parameter in pars in the form t-date-paramname}

\item{pars_temporal_base}{Base temporal parameter values, to be updated with those in pars}
}
\value{
Updated list of temporal parameters suitable for calling model_eq
}
\description{
This is a helper function since the optimizer just uses an unnamed vector of parameters.
We pass an additional vector of the names of the parameters, that can be used to determine
what parameters are in the pars vector and this overwrite the correct parameter in the 
parameter list pars_base.
}
\details{
Importantly, temporal parameters must start with t to indicate they are temporal parameters
and include the date as well as the parameter name in pars_names. For example, t-2020.05.01-rho_A
indicates to update the temporal parameters list for 1 May 2020 for rho_A. Note that the dates 
don't use dashes internally, since dashed are used to demarcate the compnents of the parameter name.

For age-varying parameters, because we can only pass a single vector to the optimizer,
these are passed as t-date-paramname_age, for example t-2020.05.01-sd_1, 
which is parsed and the appropriate parameter in the vector is updated for the specified date.
}
