% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_runner_internal.R
\name{get_update_model_days}
\alias{get_update_model_days}
\title{Gets the days to update parameters in model time}
\usage{
get_update_model_days(pars_base, pars_temporal, end_day)
}
\arguments{
\item{pars_base}{base parameters}

\item{pars_temporal}{temporal parameters}

\item{end_day}{end of simulation}
}
\value{
days to update the temporal parameters in model time
}
\description{
Gets the days to update parameters in model time
}
