% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_runner_internal.R
\name{get_update_sd_days}
\alias{get_update_sd_days}
\title{Gets the days to update sd dynamically in model time}
\usage{
get_update_sd_days(pars_base, sd_dynamics)
}
\arguments{
\item{pars_base}{base parameters}

\item{sd_dynamics}{temporal parameters for updating sd}
}
\value{
days to update dynamic sd in model time
}
\description{
Gets the days to update sd dynamically in model time
}
