% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vaccine_names.R
\name{get_vaccine_names}
\alias{get_vaccine_names}
\title{Gets the names of the vaccines}
\usage{
get_vaccine_names(params, inital_state, include_recovered_vax = TRUE)
}
\arguments{
\item{params}{parameters}

\item{inital_state}{inital state vector}

\item{include_recovered_vax}{whether to include those that are both vaccinated and recovered, otherwise just initally vaccinated}
}
\value{
names of vaccines
}
\description{
Gets the names of the vaccines which are used in the long form output.
Note that recovered from natural infection are never included.
}
