% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_state.R
\name{initial_state}
\alias{initial_state}
\title{auto-generate the initial state vector specifyin gonly S and E}
\usage{
initial_state(
  n_age,
  n_strain,
  n_vaccine,
  n_recovered,
  S_init = c(2190000 * c(0.2293, 0.4552, 0.235, 0.0805), rep(0, n_age * (n_vaccine +
    n_recovered) - n_age)),
  E_init = c(c(0, 2, 2, 0), rep(0, n_age * n_strain * (n_vaccine + n_recovered) -
    n_age))
)
}
\arguments{
\item{n_age}{number of age groups}

\item{n_strain}{number of strains}

\item{n_vaccine}{number of vaccines}

\item{n_recovered}{number of recovered}

\item{S_init}{starting susceptible population}

\item{E_init}{stating exposed population}
}
\value{
initial state vector
}
\description{
auto-generate the initial state vector for just S and E with everything else 0
note E_init will be subtracted from S_init
note the default arguments are based on the population and age distribution of King County, WA
}
