% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_runner_internal.R
\name{interpolate_value}
\alias{interpolate_value}
\title{linear interpolation at time t between value1 at time1 and value2 at time2}
\usage{
interpolate_value(t, value1, value2, time1, time2)
}
\arguments{
\item{t}{time at which to interpolate}

\item{value1}{value at time1}

\item{value2}{value at time2}

\item{time1}{time before t}

\item{time2}{time after t}
}
\value{
interpolated value
}
\description{
linear interpolation at time t between value1 at time1 and value2 at time2
}
