% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_vectorization_internal.R
\name{is_age_specfic_param}
\alias{is_age_specfic_param}
\title{Is parameter vectorized in terms of age}
\usage{
is_age_specfic_param(p)
}
\arguments{
\item{p}{parameter name}
}
\value{
TRUE if vectorized by age, FALSE otherwise
}
\description{
Note that bstar is vectorized but not age-specific
}
