% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_vectorization_internal.R
\name{is_contact_param}
\alias{is_contact_param}
\title{Does parameter change the contact matrix C}
\usage{
is_contact_param(p)
}
\arguments{
\item{p}{parameter name}
}
\value{
TRUE if changes contact matrix, FALSE otherwise
}
\description{
Does parameter change the contact matrix C
}
