% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_vectorization_internal.R
\name{is_simple_vectorized_param}
\alias{is_simple_vectorized_param}
\title{Gets if parameter p is vectorized or not}
\usage{
is_simple_vectorized_param(p)
}
\arguments{
\item{p}{parameter name}
}
\value{
TRUE if vectorized, FALSE otherwise
}
\description{
Gets if parameter p is vectorized or not
}
