% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_vectorization_internal.R
\name{is_veff_vectorized_param}
\alias{is_veff_vectorized_param}
\title{Gets if parameter p is vectorized and impacted by vaccine efficacy or not}
\usage{
is_veff_vectorized_param(p)
}
\arguments{
\item{p}{parameter name}
}
\value{
TRUE if vectorized and affected by vaccine efficacy, FALSE otherwise
}
\description{
Gets if parameter p is vectorized and impacted by vaccine efficacy or not
}
