% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_eq_internal.R
\name{model_eq}
\alias{model_eq}
\title{The core model.}
\usage{
model_eq(time, state, parameters)
}
\arguments{
\item{time}{vector of times}

\item{state}{inital state}

\item{parameters}{model parameters}
}
\value{
matrix with value of compartments at each point in time
}
\description{
Note that it should generally be run via run_model() which handles time-varying parameters. social distancing, etc. rather than directly
}
