% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving_avg.R
\name{moving_avg}
\alias{moving_avg}
\title{Calculate centered moving average}
\usage{
moving_avg(x, n = 7)
}
\arguments{
\item{x}{The vector of whcih to calculate the moving average}

\item{n}{The number of elements to include in the moving average. Must be odd.

This function calculates the centered moving average, that is, including data before and after the focal data
point. It will return a vector the same length as x with no NA's (NAs are not supprted in x). FOr the beginning and end 
of the series, a moving average of less than n is used (that is, n = 1 for the first and last data point, n = 3 for the 
2nd and 2nd to last data point).}
}
\value{
A vector of length x that is the moving average of x
}
\description{
Calculate centered moving average
}
