% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_effective_internal.R
\name{Next_Generation_Matrix}
\alias{Next_Generation_Matrix}
\title{Return the full next generation matrix}
\usage{
Next_Generation_Matrix(parameters, state, sd.mult.side)
}
\arguments{
\item{parameters}{model parameters}

\item{state}{current state}

\item{sd.mult.side}{whether sd is applied to both sides of contact matrix}
}
\value{
matrix
}
\description{
Return the full next generation matrix
}
