% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_effective_internal.R
\name{Proportion.by.state}
\alias{Proportion.by.state}
\title{Proportion by state}
\usage{
Proportion.by.state(parameters)
}
\arguments{
\item{parameters}{parameters}
}
\value{
proportions
}
\description{
Proportion by state
}
