% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_effective.R
\name{R_effective}
\alias{R_effective}
\title{R_effective}
\usage{
R_effective(parameters, state, single.Reff = T, sd.mult.side = "both")
}
\arguments{
\item{parameters}{parameters}

\item{state}{current state}

\item{single.Reff}{return a single Reffective, or strain specific}

\item{sd.mult.side}{whether sd is applied to both sides of contact matrix}
}
\value{
R effective
}
\description{
R_effective
}
