% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_vectorization_internal.R
\name{repmat}
\alias{repmat}
\title{Repeats matrix X row and column wise}
\usage{
repmat(X, rows, cols)
}
\arguments{
\item{X}{matrix}

\item{rows}{how many times to repeat}

\item{cols}{how many times to repeat}
}
\value{
repeated matrix
}
\description{
Repeats matrix X row and column wise
}
