% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{Runs the model for a specified number of days}
\usage{
run_model(
  pars_base,
  pars_temporal,
  daycount,
  init_state,
  start_day = pars_base$first_inf_day,
  calc_r_eff = TRUE
)
}
\arguments{
\item{pars_base}{base parameters}

\item{pars_temporal}{temporal parameters}

\item{daycount}{number of days to run}

\item{init_state}{initial state on start day}

\item{start_day}{(optional) to start on another model day than first_inf_day}

\item{calc_r_eff}{(optional) whether to calculate R effective}
}
\value{
matrix with daily values for all model compartments
}
\description{
This handles all the vectorization of parameters and setting the temporal parameters
}
