% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model_by_date.R
\name{run_model_by_date}
\alias{run_model_by_date}
\title{Runs the model by specifying calendar dates}
\usage{
run_model_by_date(
  pars_base,
  pars_temporal,
  init_state,
  start_date,
  end_date,
  calc_r_eff = TRUE
)
}
\arguments{
\item{pars_base}{base parameters}

\item{pars_temporal}{temporal parameters}

\item{init_state}{initial state}

\item{start_date}{start calendar date for simulation}

\item{end_date}{end calendar date for simulation}

\item{calc_r_eff}{(optional) whether to calculate R effective}
}
\value{
matrix with daily values for all model compartments
}
\description{
Runs the model by specifying calendar dates
}
