% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_data_long.R
\name{shape_data_long}
\alias{shape_data_long}
\title{Model output in long format}
\usage{
shape_data_long(ode_out, model_day0_date)
}
\arguments{
\item{ode_out}{matrix model output}

\item{model_day0_date}{calendar date for model day 0}
}
\value{
data in long format
}
\description{
Takes ode matrix output and turns in into a long data frame with columns state, strain, vac, and age.
note that strain can be empty in the S box and vac empty means not vaccinated
}
