% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_data_wide.R
\name{shape_data_wide}
\alias{shape_data_wide}
\title{Model output in wide format}
\usage{
shape_data_wide(data_long)
}
\arguments{
\item{data_long}{model output in long format}
}
\value{
model output in wide format
}
\description{
this takes the long data and puts it into wide format for comparing with kc_data
that is columns for diag, hosp, death, testpos, and testneg, each by age
where they are the daily values (and also cumulative)
note that the data is not truncated, but this can be done with a join on date
}
