% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_runner_internal.R
\name{sort_temporal_params}
\alias{sort_temporal_params}
\title{Sorts temporal parameters}
\usage{
sort_temporal_params(pars_temporal)
}
\arguments{
\item{pars_temporal}{the parameters to sort}
}
\value{
sorted parameters
}
\description{
Sorts the sublists of pars_temporal in ascending order 
and combine any lists with the same date into one list.
It is assumed that each sublist has a date element.
}
