% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_internal.R
\name{state_idx}
\alias{state_idx}
\title{Gets indices into state vector of state}
\usage{
state_idx(state, n_age, n_strain, n_vaccine, n_recovered)
}
\arguments{
\item{state}{state, see state_order}

\item{n_age}{number of age groups}

\item{n_strain}{number of strains}

\item{n_vaccine}{number of vaccines}

\item{n_recovered}{number of recovered}
}
\value{
indices
}
\description{
Gets indices into state vector of state
}
