% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_matrix_to_vector.R
\name{state_matrix_to_vector}
\alias{state_matrix_to_vector}
\title{Convert state matrix to vector}
\usage{
state_matrix_to_vector(state_matrix)
}
\arguments{
\item{state_matrix}{state in matrix format}
}
\value{
state vector
}
\description{
Utility function for manipulating states
}
