% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_vector_to_list.R
\name{state_vector_to_list}
\alias{state_vector_to_list}
\title{Convert state vector to list}
\usage{
state_vector_to_list(state_vector)
}
\arguments{
\item{state_vector}{state in vector format}
}
\value{
list
}
\description{
Utility function for manipulating states
}
