% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_vector_to_matrix.R
\name{state_vector_to_matrix}
\alias{state_vector_to_matrix}
\title{Convert state vector to matrix}
\usage{
state_vector_to_matrix(state_vector)
}
\arguments{
\item{state_vector}{state in vector format}
}
\value{
matrix
}
\description{
Utility function for manipulating states
}
