% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_runner_internal.R
\name{update_temporal_params}
\alias{update_temporal_params}
\title{Updates parameters on a certain date}
\usage{
update_temporal_params(pars, pars_to_update, sd_update_metrics = NULL)
}
\arguments{
\item{pars}{existing parameters}

\item{pars_to_update}{one item from pars_temporal which are the parameters to update}

\item{sd_update_metrics}{optionally the current metrics for determining sd dynamically}
}
\value{
updated parameters
}
\description{
Updates parameters on a certain date
}
