% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_params.R
\name{verify_params}
\alias{verify_params}
\title{Verifies parameters are internally consistent in terms of strains and vaccines, etc.}
\usage{
verify_params(params)
}
\arguments{
\item{params}{Parameters to verify}
}
\value{
none
}
\description{
this function verifies that parameters are internally consistent
for now, this verifies the dimension of other parameters based on number of strains and vaccines
Note that this should take place *before* any parameters are vectorized
}
